requirements.txt
# Agent Geronimo - Funding Opportunity Discovery System
# Core
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
playwright>=1.40.0

# Excel / Data
openpyxl>=3.1.0
pandas>=2.1.0
xlsxwriter>=3.1.0

# Config
pyyaml>=6.0.0
python-dotenv>=1.0.0

# CLI / Display
rich>=13.7.0
click>=8.1.0

# Caching / Storage
sqlite-utils>=3.35.0

# Text Processing
rapidfuzz>=3.5.0
python-dateutil>=2.8.0

# Document Reading
python-docx>=1.1.0
PyPDF2>=3.0.0
python-pptx>=0.6.21
markdown>=3.5.0

# Rate Limiting
ratelimit>=2.2.1

# RSS
feedparser>=6.0.0

# Optional: semantic similarity
# sentence-transformers>=2.2.0
# numpy>=1.24.0
